requirements.txt
fastapi
uvicorn[standard]
pydantic
numpy
faiss-cpu
sentence-transformers
uvicorn
requests
python-multipart
aiofiles
sqlmodel